Move `ModuleName` into a new file
